+++ js/privatebin.js
@@ -3932,7 +3932,26 @@ jQuery.PrivateBin = (function($, RawDeflate) {
          */
         function setLanguage(event)
         {
-            document.cookie = 'lang=' + $(event.target).data('lang') + '; SameSite=Lax; Secure';
+            let lang = $(event.target).data('lang') || event.target.value;
+
+            document.cookie = 'lang=' + lang + '; SameSite=Lax; Secure';
+            window.location.reload();
+            event.preventDefault();
+        }
+
+        /**
+         * save the template in a cookie and reloads the page
+         *
+         * @name TopNav.setTemplate
+         * @private
+         * @function
+         * @param {Event} event
+         */
+        function setTemplate(event)
+        {
+            let template = $(event.target).data('template') || event.target.value;
+
+            document.cookie = 'template=' + template + '; SameSite=Lax; Secure';
             window.location.reload();
             event.preventDefault();
         }
@@ -4624,7 +4643,12 @@ jQuery.PrivateBin = (function($, RawDeflate) {
             // bootstrap template drop down
             $('#language ul.dropdown-menu li a').click(setLanguage);
             // page template drop down
-            $('#language select option').click(setLanguage);
+            $('#language select').change(setLanguage);
+
+            // bootstrap template drop down
+            $('#template ul.dropdown-menu li a').click(setTemplate);
+            // page template drop down
+            $('#template select').change(setTemplate);
 
             // bind events
             $burnAfterReading.change(changeBurnAfterReading);