+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # PrivateBin version history
 
 ## 1.7.7 (not yet released)
+* ADDED: Switching templates using the web ui (#13)
 * CHANGED: Upgrading libraries to: ip-lib 1.20.0
 
 ## 1.7.6 (2025-02-01)